+++ tensorflow/core/util/tensor_slice_writer.h
@@ -65,7 +65,7 @@ class TensorSliceWriter {
   // Allocate "num_elements" elements in "ss" and save the data in "data"
   // there.
   template <typename T>
-  static Status SaveData(const T* data, int num_elements, SavedSlice* ss);
+  static Status SaveData(const T* data, int64 num_elements, SavedSlice* ss);
 
   static size_t MaxBytesPerElement(DataType dt);
 
@@ -162,7 +162,7 @@ Status TensorSliceWriter::Add(const string& name, const TensorShape& shape,
 }
 
 template <typename T>
-Status TensorSliceWriter::SaveData(const T* data, int num_elements,
+Status TensorSliceWriter::SaveData(const T* data, int64 num_elements,
                                    SavedSlice* ss) {
   size_t size_bound =
       ss->ByteSize() + kTensorProtoHeaderBytes +
@@ -179,7 +179,7 @@ Status TensorSliceWriter::SaveData(const T* data, int num_elements,
 }
 
 template <>
-Status TensorSliceWriter::SaveData(const string* data, int num_elements,
+Status TensorSliceWriter::SaveData(const string* data, int64 num_elements,
                                    SavedSlice* ss);
 
 // Create a table builder that will write to "filename" in
